Readme update and bug fix
Date picker was malfunctioning after a fresh install with yarn. Due to a dploy bug we can not have yarn.lock file present in the build. So I removed the ^ (carrot charactors) from all packages in package.json so the package manager would only install those versions in the package-lock.json. Which solved the issue. this is a bandaid fix to a larger problem with package incompatibility and deploy bugs